Fix #123 - request paused on exit
It appears that subsequent operations which have the effect of pausing the
request are already queued for this event frame. This rux adds setImmediate to
push stream resumption to the next frame.

In the future, we should run tests with a small payload AND one that exceeds the
maximum TCP packet size limit and node internal buffer. This would help identify
this kind of bug.